added local voting protocol

+++ src/main/kotlin/AverageAgent.kt
@@ -6,7 +6,7 @@ class AverageAgent: Agent() {
         println(name)
 
         if (arguments != null && arguments.isNotEmpty()) {
-            val number = arguments[0].toString().toInt()
+            val number = arguments[0].toString().toDouble()
             val peers = arguments
                 .slice(1 until arguments.size)
                 .map { AID(it.toString(), false) }
